+++ ICD.Connect.Devices/Devices/Simpl/AbstractSimplDeviceBase.cs
@@ -42,7 +42,7 @@ namespace ICD.Connect.Devices.Simpl
 
 				m_IsOnline = value;
 
-				Log(eSeverity.Informational, "Online status changed to {0}", IsOnline);
+				Logger.Set("Online", eSeverity.Informational, IsOnline);
 
 				OnIsOnlineStateChanged.Raise(this, new DeviceBaseOnlineStateApiEventArgs(IsOnline));
 			}
@@ -61,7 +61,7 @@ namespace ICD.Connect.Devices.Simpl
 
 				m_ControlsAvailable = value;
 
-				Log(eSeverity.Informational, "Controls Available changed to {0}", ControlsAvailable);
+				Logger.Set("Controls Available", eSeverity.Informational, ControlsAvailable);
 
 				OnControlsAvailableChanged.Raise(this, new DeviceBaseControlsAvailableApiEventArgs(ControlsAvailable));
 			}