+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
  - Added MockDevice interfaces/classes to allow Mock devices to be set online/offline from console and be configured to start offline from settings
+
+### Changed
+ - Using new logging context
 
 ## [10.0.0] - 2020-03-20
 ### Added